LayoutTests/media/track/track-webvtt-tc000-empty.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>            
 
            function trackLoaded()
            {
                findMediaElement();
                testExpected("video.textTracks[0].cues.length", "0");
                
                endTest();
            }

        </script>
    </head>
    <body>
        <p>Tests that an empty file is not recognized as a WebVTT file.</p>
        <video>
            <track src="captions-webvtt/tc000-empty.vtt" onload="trackLoaded()" default>
        </video>
    </body>
</html>
